test: re-enable async replication acceptance tests

test: re-enable async replication acceptance tests
